+++ chameleonultragui/lib/bridge/dfu.dart
@@ -3,6 +3,7 @@ import 'dart:typed_data';
 import 'dart:async';
 import 'package:chameleonultragui/helpers/general.dart';
 import 'package:chameleonultragui/connector/serial_abstract.dart';
+import 'package:flutter/foundation.dart';
 import 'package:logger/logger.dart';
 import 'dart:math';
 
@@ -205,6 +206,7 @@ class DFUCommunicator {
   Future<dynamic> selectObject(int objectType) async {
     var response = (await sendCmd(DFUCommand.readObject,
         Uint8List.fromList([objectType, 0x00, 0x00, 0x00])))!;
+    print(response.buffer.lengthInBytes);
     var maxSize = ByteData.view(response.buffer).getUint32(0, Endian.little);
     var offset = ByteData.view(response.buffer).getUint32(4, Endian.little);
     var crc = ByteData.view(response.buffer).getUint32(8, Endian.little);
@@ -243,7 +245,12 @@ class DFUCommunicator {
   }
 
   Future<Map<String, int>> calculateChecksum() async {
+    print('checksum');
     var response = await sendCmd(DFUCommand.calcChecSum, Uint8List(0));
+    if (response!.buffer.lengthInBytes < 8) {
+      print('checksum resend due to low bytes');
+      response = await sendCmd(DFUCommand.calcChecSum, Uint8List(0));
+    }
     var offset = ByteData.view(response!.buffer).getUint32(0, Endian.little);
     var crc = ByteData.view(response.buffer).getUint32(4, Endian.little);
 
@@ -259,7 +266,7 @@ class DFUCommunicator {
     for (var offset = 0; offset < firmwareBytes.length; offset += length) {
       var tries = 0;
       var crcBackup = crc;
-      for (; tries < ((Platform.isIOS) ? 50 : 10); tries++) {
+      for (; tries < ((!kIsWeb && Platform.isIOS) ? 50 : 10); tries++) {
         await createObject(
             objectType, min(firmwareBytes.length - offset, length));
 
@@ -283,7 +290,7 @@ class DFUCommunicator {
         break;
       }
 
-      if (tries == ((Platform.isIOS) ? 50 : 10)) {
+      if (tries == ((!kIsWeb && Platform.isIOS) ? 50 : 10)) {
         throw ("Unable to recover from DFU");
       }
     }
@@ -325,11 +332,14 @@ class DFUCommunicator {
       offset += toTransmit.length;
       crc = calculateCRC32(toTransmit, crc) & 0xFFFFFFFF;
       currentPrn++;
+      print(currentPrn);
       if (currentPrn == prn) {
         await asyncSleep(1);
         response = await calculateChecksum();
         validateCrc();
         currentPrn = 0;
+      } else {
+        print('${currentPrn} is not yet ${prn}');
       }
     }
 
@@ -348,7 +358,7 @@ class DFUCommunicator {
       offsetSize = 20;
     }
 
-    if (Platform.isWindows || Platform.isMacOS || isBLE) {
+    if (kIsWeb || Platform.isWindows || Platform.isMacOS || isBLE) {
       for (var offset = 0; offset < packet.length; offset += offsetSize) {
         await _serialInstance!.write(
             packet.sublist(
@@ -356,7 +366,7 @@ class DFUCommunicator {
             firmware: true);
       }
 
-      if (isBLE && (Platform.isIOS || Platform.isMacOS)) {
+      if (kIsWeb || isBLE && (Platform.isIOS || Platform.isMacOS)) {
         await asyncSleep(250);
       }
     } else {
