+++ chameleonultragui/lib/connector/serial_abstract.dart
@@ -26,6 +26,7 @@ abstract class AbstractSerial {
   bool isDFU = false;
   bool pendingConnection = false;
   String portName = "None";
+  String name = "Abstract";
   ConnectionType connectionType = ConnectionType.none;
   dynamic messageCallback;
