+++ chameleonultragui/lib/gui/menu/chameleon_settings.dart
@@ -103,6 +103,7 @@ class ChameleonSettingsState extends State<ChameleonSettings> {
                                 await flashFirmware(appState,
                                     scaffoldMessenger: scaffoldMessenger);
                               } catch (e) {
+                                appState.log!.e(e);
                                 scaffoldMessenger.hideCurrentSnackBar();
                                 snackBar = SnackBar(
                                   content: Text(
@@ -133,6 +134,7 @@ class ChameleonSettingsState extends State<ChameleonSettings> {
                                 await flashFirmwareZip(appState,
                                     scaffoldMessenger: scaffoldMessenger);
                               } catch (e) {
+                                appState.log!.e(e);
                                 scaffoldMessenger.hideCurrentSnackBar();
                                 var snackBar = SnackBar(
                                   content: Text(
