+++ chameleonultragui/lib/helpers/github.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 import 'dart:typed_data';
 import 'package:chameleonultragui/connector/serial_abstract.dart';
+import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
 
 List<Map<String, String>> developers = [
@@ -21,12 +22,6 @@ List<Map<String, String>> developers = [
     'description': 'Developer',
     'avatarUrl': 'https://avatars.githubusercontent.com/u/13242738',
     'username': 'augustozanellato'
-  },
-  {
-    'name': 'Thomas Cannon',
-    'description': 'Apple maintainer',
-    'avatarUrl': 'https://avatars.githubusercontent.com/u/1297160',
-    'username': 'thomascannon'
   },
   {
     'name': 'Akisame',
@@ -99,6 +94,11 @@ Future<Uint8List> fetchFirmwareFromReleases(ChameleonDevice device) async {
     for (var file in releases[0]["assets"]) {
       if (file["name"] ==
           "${(device == ChameleonDevice.ultra) ? "ultra" : "lite"}-dfu-app.zip") {
+        if (kIsWeb) {
+          file["browser_download_url"] =
+              "https://api.codetabs.com/v1/proxy/?quest=${file["browser_download_url"]}";
+        }
+
         content = await http.readBytes(Uri.parse(file["browser_download_url"]));
         break;
       }
