+++ chameleonultragui/lib/main.dart
@@ -1,18 +1,13 @@
 import 'dart:io';
 import 'package:chameleonultragui/bridge/chameleon.dart';
 import 'package:chameleonultragui/connector/serial_abstract.dart';
-import 'package:chameleonultragui/connector/serial_android.dart';
-import 'package:chameleonultragui/connector/serial_ble.dart';
 import 'package:chameleonultragui/connector/serial_emulator.dart';
-import 'package:chameleonultragui/connector/serial_macos.dart';
 import 'package:chameleonultragui/helpers/font.dart';
 import 'package:chameleonultragui/helpers/general.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:wakelock_plus/wakelock_plus.dart';
-
-import 'connector/serial_native.dart';
 
 // Page imports
 import 'package:chameleonultragui/gui/page/home.dart';
@@ -34,6 +29,13 @@ import 'package:chameleonultragui/sharedprefsprovider.dart';
 
 // Logger
 import 'package:logger/logger.dart';
+
+import 'package:chameleonultragui/connector/serial_stub.dart'
+    if (dart.library.js_util) 'package:chameleonultragui/connector/serial_web.dart'
+    if (Platform.isAndroid) 'package:chameleonultragui/connector/serial_android.dart'
+    if (Platform.isMacOS) 'package:chameleonultragui/connector/serial_macos.dart'
+    if (Platform.isIOS) 'package:chameleonultragui/connector/serial_ble.dart'
+    if (dart.library.io) 'package:chameleonultragui/connector/serial_native.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -126,19 +128,7 @@ class _MainPageState extends State<MainPage> {
       return EmulatorSerial(log: appState.log!);
     }
 
-    if (Platform.isMacOS) {
-      return MacOSSerial(log: appState.log!);
-    }
-
-    if (Platform.isAndroid) {
-      return AndroidSerial(log: appState.log!);
-    }
-
-    if (Platform.isIOS) {
-      return BLESerial(log: appState.log!);
-    }
-
-    return NativeSerial(log: appState.log!);
+    return SerialAdapter(log: appState.log!);
   }
 
   Logger getLogger(ChameleonGUIState appState) {
@@ -173,6 +163,8 @@ class _MainPageState extends State<MainPage> {
     }
 
     appState.devMode = appState.sharedPreferencesProvider.isDebugMode();
+
+    print(appState.connector!.connected);
 
     Widget page; // Set Page
     if (!appState.connector!.connected &&
