+++ chameleonultragui/lib/recovery/recovery_stub.dart
@@ -1,0 +1,26 @@
+import 'package:chameleonultragui/recovery/definitions.dart';
+
+Future<List<int>> darkside(DarksideDart darkside) async {
+  return [];
+}
+
+Future<List<int>> nested(NestedDart nested) async {
+  return [];
+}
+
+Future<List<int>> hardNested(HardNestedDart nested) async {
+  return [];
+}
+
+Future<List<int>> staticNested(StaticNestedDart nested) async {
+  return [];
+}
+
+Future<List<int>> staticEncryptedNested(
+    StaticEncryptedNestedDart nested) async {
+  return [];
+}
+
+Future<List<int>> mfkey32(Mfkey32Dart mfkey) async {
+  return [];
+}
