+++ chameleonultragui/pubspec.yaml
@@ -19,7 +19,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.2.0+0
 
 environment:
-  sdk: ">3.0.0"
+  sdk: ">3.4.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -71,6 +71,8 @@ dependencies:
   wakelock_plus: ^1.2.4
   qr_flutter: ^4.1.0
   mobile_scanner: ^7.0.1
+  serial: ^0.0.7+1
+  web: ^1.1.1
 
 dev_dependencies:
   flutter_test:
